requirements.txt
lxml==4.2.5
Pillow==5.2.0